Added binding events and realtime handling of state changes via the dummy panel (wxSliderPanel -> wxDummySliderPanel and renamed wxSubSliderPanel -> wxSliderPanel)
