+++ OutfitStudio.vcxproj
@@ -765,9 +765,9 @@
     <ClInclude Include="src\render\GLShader.h" />
     <ClInclude Include="src\render\GLSurface.h" />
     <ClInclude Include="src\ui\wxBrushSettingsPopup.h" />
-    <ClInclude Include="src\ui\wxSliderPanel.h" />
+    <ClInclude Include="src\ui\wxDummySliderPanel.h" />
     <ClInclude Include="src\ui\wxStateButton.h" />
-    <ClInclude Include="src\ui\wxSubSliderPanel.h" />
+    <ClInclude Include="src\ui\wxSliderPanel.h" />
     <ClInclude Include="src\utils\AABBTree.h" />
     <ClInclude Include="src\utils\ConfigurationManager.h" />
     <ClInclude Include="src\utils\ConfigDialogUtil.h" />
@@ -839,9 +839,9 @@
     <ClCompile Include="src\render\GLShader.cpp" />
     <ClCompile Include="src\render\GLSurface.cpp" />
     <ClCompile Include="src\ui\wxBrushSettingsPopup.cpp" />
-    <ClCompile Include="src\ui\wxSliderPanel.cpp" />
+    <ClCompile Include="src\ui\wxDummySliderPanel.cpp" />
     <ClCompile Include="src\ui\wxStateButton.cpp" />
-    <ClCompile Include="src\ui\wxSubSliderPanel.cpp" />
+    <ClCompile Include="src\ui\wxSliderPanel.cpp" />
     <ClCompile Include="src\utils\AABBTree.cpp" />
     <ClCompile Include="src\utils\ConfigurationManager.cpp" />
     <ClCompile Include="src\utils\Log.cpp" />
