+++ OutfitStudio.vcxproj.filters
@@ -1753,7 +1753,7 @@
     <ClInclude Include="src\ui\wxBrushSettingsPopup.h">
       <Filter>UI</Filter>
     </ClInclude>
-    <ClInclude Include="src\ui\wxSliderPanel.h">
+    <ClInclude Include="src\ui\wxDummySliderPanel.h">
       <Filter>UI</Filter>
     </ClInclude>
     <ClInclude Include="src\program\ConvertBodyReferenceDialog.h">
@@ -1765,7 +1765,7 @@
     <ClInclude Include="src\program\SliderDataImportDialog.h">
       <Filter>Program</Filter>
     </ClInclude>
-    <ClInclude Include="src\ui\wxSubSliderPanel.h" />
+    <ClInclude Include="src\ui\wxSliderPanel.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="lib\TinyXML-2\tinyxml2.cpp">
@@ -1969,7 +1969,7 @@
     <ClCompile Include="src\ui\wxBrushSettingsPopup.cpp">
       <Filter>UI</Filter>
     </ClCompile>
-    <ClCompile Include="src\ui\wxSliderPanel.cpp">
+    <ClCompile Include="src\ui\wxDummySliderPanel.cpp">
       <Filter>UI</Filter>
     </ClCompile>
     <ClCompile Include="src\program\ConvertBodyReferenceDialog.cpp">
@@ -1978,7 +1978,7 @@
     <ClCompile Include="src\program\SliderDataImportDialog.cpp">
       <Filter>Program</Filter>
     </ClCompile>
-    <ClCompile Include="src\ui\wxSubSliderPanel.cpp" />
+    <ClCompile Include="src\ui\wxSliderPanel.cpp" />
   </ItemGroup>
   <ItemGroup>
     <Xml Include="Config.xml">
